quick save hp gui wip

+++ Assets/Scripts/Enemy/Enemy.cs
@@ -6,15 +6,15 @@ public class Enemy : MonoBehaviour
 {
     private GameObject map;
 
-    private int health = 10;
-    private int damage = 1;
+    private float health = 10;
+    private float damage = 1;
 
     private void Start()
     {
         map = GameObject.FindWithTag("Map");
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         health -= damage;
 
@@ -24,7 +24,7 @@ public class Enemy : MonoBehaviour
         }
     }
 
-    public int getDamage()
+    public float getDamage()
     {
         return damage;
     }
